ai_project_proposal_writer/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.136.0",
    "google-generativeai>=0.8.6",
    "groq>=1.2.0",
    "httpx>=0.28.1",
    "litellm>=1.83.0",
    "openai-agents>=0.14.4",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "together>=2.9.0",
    "uvicorn>=0.45.0",
]
